# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/containers": {
      "get": {
        "description": "list all containers",
        "produces": [
          "application/json"
        ],
        "tags": [
          "containers"
        ],
        "summary": "list all containers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.ContainerListResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.ErrorResponse"
            }
          }
        }
      }
    },
    "/containers/{containerId}/logs": {
      "get": {
        "description": "get container logs",
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "containers"
        ],
        "summary": "get container logs",
        "parameters": [
          {
            "type": "string",
            "description": "container id",
            "name": "containerId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/containers/{containerId}/{command}": {
      "post": {
        "description": "execute a command on a container",
        "produces": [
          "application/json"
        ],
        "tags": [
          "containers"
        ],
        "summary": "execute a command on a container",
        "parameters": [
          {
            "type": "string",
            "description": "container id",
            "name": "containerId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "command to execute (start/stop/restart)",
            "name": "command",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.OkResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.ErrorResponse"
            }
          }
        }
      }
    },
    "/openmower/call/{command}": {
      "post": {
        "description": "call a service",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "openmower"
        ],
        "summary": "call a service",
        "parameters": [
          {
            "type": "string",
            "description": "command to call, could be: mower_start, mower_home, mower_s1, mower_s2, emergency, mow",
            "name": "command",
            "in": "path",
            "required": true
          },
          {
            "description": "request body",
            "name": "CallReq",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.OkResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.ErrorResponse"
            }
          }
        }
      }
    },
    "/openmower/subscribe/{topic}": {
      "get": {
        "description": "subscribe to a topic",
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "openmower"
        ],
        "summary": "subscribe to a topic",
        "parameters": [
          {
            "type": "string",
            "description": "topic to subscribe to, could be: diagnostics, status, gps, imu, ticks",
            "name": "topic",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/settings": {
      "get": {
        "description": "returns a JSON object with the settings",
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "summary": "returns a JSON object with the settings",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.GetSettingsResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "saves the settings to the mower_config.sh file",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "settings"
        ],
        "summary": "saves the settings to the mower_config.sh file",
        "parameters": [
          {
            "description": "settings",
            "name": "settings",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/api.OkResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/api.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "api.Container": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "names": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "state": {
          "type": "string"
        }
      }
    },
    "api.ContainerListResponse": {
      "type": "object",
      "properties": {
        "containers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/api.Container"
          }
        }
      }
    },
    "api.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "api.GetSettingsResponse": {
      "type": "object",
      "properties": {
        "settings": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "api.OkResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "string"
        }
      }
    }
  }
}